break down htnn/pkg
Now it is two packages:
* pkg: for the outside developers
* internal: used internally
