+++ plugins/tests/integration/control_plane/control_plane.go
@@ -42,10 +42,10 @@ import (
 	"google.golang.org/grpc"
 	"google.golang.org/protobuf/types/known/structpb"
 
+	"mosn.io/htnn/internal/proto"
 	"mosn.io/htnn/pkg/filtermanager"
 	"mosn.io/htnn/pkg/filtermanager/model"
 	"mosn.io/htnn/pkg/log"
-	"mosn.io/htnn/pkg/proto"
 	"mosn.io/htnn/plugins/tests/integration/data_plane"
 )
